% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFunctions.R
\name{binarySearchLower}
\alias{binarySearchLower}
\title{binarySearchLower}
\usage{
binarySearchLower(lowerLimit, upperLimit, numberInspections)
}
\arguments{
\item{lowerLimit}{The lowerlimit of the search passed to the function}

\item{upperLimit}{The upperLimit of the search passed to the function}

\item{numberInspections}{The current number of inspection}
}
\description{
A general search function to find the root for the upper search
}
