% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFunctions.R
\name{binarySearchV}
\alias{binarySearchV}
\title{binarySearchV}
\usage{
binarySearchV(lowerLimit, upperLimit, numberInspections)
}
\arguments{
\item{lowerLimit}{The lowerlimit of the search passed to the function}

\item{upperLimit}{The upperLimit of the search passed to the function}

\item{numberInspections}{The current number of inspection}
}
\description{
THe binary search v function is a bsic search function to find the root for V
}
