% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerIntegrandDensity.R
\name{densityArray}
\alias{densityArray}
\title{densityArray}
\usage{
densityArray(get_power, numberInspections)
}
\arguments{
\item{get_power}{Get_power is an argument to determine the power of the trial}

\item{numberInspections}{Number of inspections in the current analysis}
}
\description{
This function is used to compute the densities for each analysis in the trial in order to recursively find the stopping boundaries
}
