% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Generate Data Function}
\usage{
generate.data(n1, mean.bin, mean.t, sigma, rho)
}
\arguments{
\item{n1}{n1 is the number of people recruited overall in the study}

\item{mean.bin}{Mean.bin is the mean of the binary endpoint}

\item{mean.t}{Mean.t is the mean of the normally distributed endpoint}

\item{sigma}{Sigma is the standard deviation of the endpoints, which is assumed to be equal}

\item{rho}{Rho is the correlation between the endpoints}
}
\description{
This function is used to generate data for use in a simulation of the seamless trails with different endpoints
}
