% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.z.v.full.R
\name{get.z.v.bin}
\alias{get.z.v.bin}
\title{get.z.v.bin}
\usage{
get.z.v.bin(full.data)
}
\arguments{
\item{full.data}{Data containing the binomally distributed endpoint and the treatment assignments}
}
\description{
This function calculates the Z and V test statistics for a binomially distributed endpoint
}
