% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.z.v.full.R
\name{get.z.v.full}
\alias{get.z.v.full}
\title{get.z.v.full function}
\usage{
get.z.v.full(full.data, inspection, dist = "normal")
}
\arguments{
\item{inspection}{The current inspection in the trial}

\item{dist}{Dist the distribution of the endpoint at this inspection}

\item{data}{Data is data containing all the interim analysis data. In the form of a long-term endpoint followed by a short-term endpoint column.}
}
\description{
This function is used to calculate the Z and V test statistics
}
