% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.z.v.full.R
\name{get.z.v.norm}
\alias{get.z.v.norm}
\title{get.z.v.norm function}
\usage{
get.z.v.norm(full.data)
}
\arguments{
\item{data}{The data containing a noramlly distributed endpoint and treatment assignments}
}
\description{
This function computes the Z and V test statistics for a normally distributed endpoint
}
