% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdensityandpower.R
\name{get_density_first}
\alias{get_density_first}
\title{get_density_first}
\usage{
get_density_first(ntmt, theta, v, z)
}
\arguments{
\item{ntmt}{Ntmt is the number of treatments}

\item{theta}{Theta is the treatment effect variable}

\item{v}{V is the value of the fisher information statistic}

\item{z}{Z is the current value for the Efficient Score statistic}
}
\description{
This function gets the density of the test statistic z at the first analysis
}
