% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdensityandpower.R
\name{get_integrand}
\alias{get_integrand}
\title{get_integrand}
\usage{
get_integrand(k, theta, v, z, x)
}
\arguments{
\item{k}{K is the number of treatments used in this trial before the trial}

\item{theta}{Theta is the treament effect}

\item{v}{V is the current fisher's information}

\item{z}{Z is the current}

\item{x}{X is used in the integrand}
}
\description{
This is a function that returns the integrand for the
}
