% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdensityandpower.R
\name{get_integrand_power}
\alias{get_integrand_power}
\title{get_integrand_power}
\usage{
get_integrand_power(k, theta, v, z, x)
}
\arguments{
\item{theta}{Theta is the treatment effect}

\item{v}{V is the Fisher information test statistic}

\item{z}{Z is the current Efficient Score test statistic}

\item{x}{X is the value used to find power}

\item{ntmt}{Ntmt is the number of treatments}
}
\description{
This function gets the power for the first test statistic
}
