% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdensityandpower.R
\name{get_power_first}
\alias{get_power_first}
\title{get_power_first}
\usage{
get_power_first(ntmt, theta, v, z)
}
\arguments{
\item{ntmt}{Ntmt is the number of treatments}

\item{theta}{Theta is the treatment effect}

\item{v}{V is the Fisher information test statistic}

\item{z}{Z is the current Efficient Score test statistic}
}
\description{
This function gets the power for the first test statistic
}
