% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerIntegrandDensity.R
\name{getErrorRates}
\alias{getErrorRates}
\title{getErrorRates}
\usage{
getErrorRates(get_power, numberInspections, print)
}
\arguments{
\item{get_power}{Get_power determines whether or not this is returning the power of the trial}

\item{numberInspections}{The number of inspections in the trial}

\item{print}{Print is an option to print the error rates or not}
}
\description{
This function gets the error rates for trial
}
