% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARfunctions.R
\name{getTreat}
\alias{getTreat}
\title{Function for getting the treatment assignments}
\usage{
getTreat(covValues, best = 0, tr = NULL, n.trt, design)
}
\arguments{
\item{covValues}{Covariate Values for the number of patients in the trial}

\item{best}{Best is the selected best treatment after the first analysis}

\item{Tr}{Treatment assignments passed back to the treatment assignment function after the first interim analysis}

\item{Design}{Design is the variable determing whether there to use Pocock's design or Stratified Permuted Block Randomization}
}
\description{
Get treamtments points either to Pocock's design or Stratified Permuted Block Design
}
