% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerIntegrandDensity.R
\name{integrateNormalDistributionDensityFunctionDiscrete}
\alias{integrateNormalDistributionDensityFunctionDiscrete}
\title{integrateNormalDistributionDensityFunctionDiscrete}
\usage{
integrateNormalDistributionDensityFunctionDiscrete(inspection)
}
\arguments{
\item{inspection}{Inspection is the current inspection in the trial}
}
\description{
This function is used to compute the integral in the probCrossUpperDiscrete function
}
