% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFunctions.R
\name{lower_search_fn}
\alias{lower_search_fn}
\title{Lower search function}
\usage{
lower_search_fn(lowerBoundary, numberInspections)
}
\arguments{
\item{lowerBoundary}{The upperBoundary passed to upper_search_fn}

\item{numberInspections}{The current number of numberInspections}
}
\description{
This is a function to find the upper boundary
}
