% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerIntegrandDensity.R
\name{probCrossLowerDiscrete}
\alias{probCrossLowerDiscrete}
\title{probCrossLowerDiscrete}
\usage{
probCrossLowerDiscrete(get_power, inspection)
}
\arguments{
\item{get_power}{Get_power determines the power in the trial}

\item{inspection}{Inpsection is current inspection in the trial}
}
\description{
This function is used to compute the probability of crossing the lower boundary
}
