% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerIntegrandDensity.R
\name{probCrossUpperDiscrete}
\alias{probCrossUpperDiscrete}
\title{probCrossUpperDiscrete}
\usage{
probCrossUpperDiscrete(get_power, inspection)
}
\arguments{
\item{get_power}{get_power is the arugment to calculate the power in the trial}

\item{inspection}{Inspection is the current inspection in the analysis}
}
\description{
This function is used to determine the probability of crossing the upper boundary
}
