% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFunctions.R
\name{simbin}
\alias{simbin}
\title{simbin function}
\usage{
simbin(mean.s, trts, n.s, treat)
}
\arguments{
\item{mean.s}{mean.s is the mean of the binary endpoint}

\item{trts}{Trts is the vector of possible treatments}

\item{n.s}{N.s is the frequency of each treatment}
}
\description{
This function is used to simulate binary endpoint data
}
