% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFunctions.R
\name{simnormal}
\alias{simnormal}
\title{Sim normal function}
\usage{
simnormal(mean.s, sigma, trts, n.s, tau1, tau2, get_power = 0, covValues,
  treat)
}
\arguments{
\item{mean.s}{Mean.s is the mean of the outcome varaible}

\item{sigma}{Sigma is the variance}

\item{trts}{Trts is the possible treatment assignments}

\item{n.s}{N.s is the frequency of each treatment}

\item{tau1}{Tau1 is the weight assigned to the first covariate value}

\item{tau2}{Tau2 is the weight assigned to the second covariate value}

\item{get_power}{Get_power determines whether or not the power is being calculated in this simulation}
}
\description{
Function to simulate normally distributed data based on given treatment groups
}
