% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFunctions.R
\name{simulatedata.car}
\alias{simulatedata.car}
\title{Simulate Data For CAR Function}
\usage{
simulatedata.car(mean.s, sigma = 1, tau1 = 1, tau2 = 1, treat,
  covValues, data = NULL, dist, inspection)
}
\arguments{
\item{mean.s}{Mean of the short term response}

\item{sigma}{sigma in bivariate normal distribution}

\item{tau1}{Tau1 is the weight on the first covariate}

\item{tau2}{Tau2 is the weight on the second covariate}

\item{treat}{List containing the treatment assignments and the base matrix}

\item{covValues}{covValues is the dummy variable matrix from the covariate values}

\item{data}{Is the data from the previous analysis if applicable. The default is NULL}
}
\description{
Function to simulate data with treatment assignments from Covariate Adaptive Randomization Scheme
}
