% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFunctions.R
\name{upper_search_fn}
\alias{upper_search_fn}
\title{Upper search function}
\usage{
upper_search_fn(upperBoundary, numberInspections)
}
\arguments{
\item{upperBoundary}{The upperBoundary passed to upper_search_fn}

\item{numberInspections}{The current number of numberInspections}
}
\description{
This is a function to find the upper boundary
}
