% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFunctions.R
\name{v_search_fn}
\alias{v_search_fn}
\title{V Search function}
\usage{
v_search_fn(vmax)
}
\arguments{
\item{vmax}{Vmax is the max value passed to this function}
}
\description{
Function to search for solution to V distribution
}
